Extract ParsedJavadoc and BlockTag into their own files for clarity
